Need to document more, but.
Task 1 to 5 now works. I didn't commit much from start to that point. But I still have some cleaning up to do.

So The commits will come with that.

task 6 fails on 8 tasks. So I will have to figure out a way to sort them, other than that the overall progression is good.